requirements.txt
streamlit==1.27.0
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.21.3
langchain==0.0.300
speechrecognition==3.8.1
gtts==2.2.3
yfinance==0.2.28
alpha-vantage==2.3.1
pyaudio==0.2.1
pandas
numpy
openai==1.3.7
langchain-openai==0.0.2
python-multipart==0.0.6
pydantic==2.5.0
pyttsx3==2.90
sounddevice==0.4.6
scipy==1.11.4
crewai==0.1.0
ollama==0.1.7
litellm==1.17.0
chromadb==0.4.22
aiofiles
asyncio
httpx
langchain-community
